+++ scripts/macros.js
@@ -2,4 +2,7 @@ export {acidSplash} from './macros/2024/spells/acidSplash.js';
 export {aid} from './macros/2024/spells/aid.js';
 export {alterSelf} from './macros/2024/spells/alterSelf.js';
 export {animalFriendship} from './macros/2024/spells/animalFriendship.js';
+export {animalMessenger} from './macros/2024/spells/animalMessenger.js';
+export {animateDead} from './macros/2024/spells/animateDead.js';
+export {antilifeShell, antilifeShellOtherMove} from './macros/2024/spells/antilifeShell.js';
 export {armorOfAgathys} from './macros/2024/spells/armorOfAgathys.js';